pyproject.toml
[tool.poetry]
name = "socket-vdash"
version = "0.1.0"
description = "Socket Security Value Dashboard - Grafana integration for Socket.dev metrics"
authors = ["Jenna"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
schedule = "^1.2.1"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.0"
black = "^24.0.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
